Rename src as homeworks directory and crate __init__.py (#27)
Rename src as homeworks directory and crate __init__.py
