+++ homeworks/communication_manager.py
