+++ homeworks/logging_console.py
