+++ homeworks/store_manager.py
