+++ initialize_infra.py
@@ -8,9 +8,9 @@
 """
 
 import sys
-import src.logging_console as logging_console
-from src.store_manager import Store_manager
-from src.communication_manager import Communication_manager
+from homeworks import logging_console
+from homeworks.communication_manager import Communication_manager
+from homeworks.store_manager import Store_manager
 
 log = logging_console.getLogger("homeworks")
