--- tests/ca.pem
@@ -1,25 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIEQTCCAqmgAwIBAgIUQzdtBSMgNh23QDuj4wY5QBjCwi8wDQYJKoZIhvcNAQEM
-BQAwOjE4MDYGA1UEAwwvMjAwMWE5NTEtZjllNC00MmQ1LTk1OTgtMjZhZmRlNzFh
-ODA2IFByb2plY3QgQ0EwHhcNMjEwNDIwMTI0NjA5WhcNMzEwNDE4MTI0NjA5WjA6
-MTgwNgYDVQQDDC8yMDAxYTk1MS1mOWU0LTQyZDUtOTU5OC0yNmFmZGU3MWE4MDYg
-UHJvamVjdCBDQTCCAaIwDQYJKoZIhvcNAQEBBQADggGPADCCAYoCggGBAO4YzXPb
-ccfjFBdorPmCu+gBqZQZXziRNYf/DUm21BIc95Mkrb2uzDOvUgOXeTAPZkU0B41t
-Wypom+yM/PuCtV8Zq9lSED1bQNSD6IAIFzo2dF+gtBWYIk+9TncNK4xzKEo9J+sV
-SX1Bk3tiErQ4n6DDAiPeeYw5Y5DyDrfMYIDvxxfbUCFFGg6MKRrfIHQeTR4IeH+s
-tk4b24gipV2OMOshXNeODVbgPrMZXKUKyQn7jYqTsgLlKLs7GT2NunsX7vMddqWj
-GgPguHEWQx2vEIcL7nyv+QZQ1lZw0XDsoJ/oG0bdm+QQd/d71xBpSry2AjZB13W8
-4QKMIJZnjyzFBdrsnoutj7dLt7o3Th/SQDXBjqmKy7qfqy3O9G+vnFpw+Aypu0U3
-DYz2FCpQd/7MRamGbPyryiTpqPBEfWe6Ug6y018dgwHuIOR4VzmimcbHHy41+kKA
-PegVOOhtOMXu2hgvHXkkK/OrdaEw1MywyKrlIoGdQ1RPjD55+jw9k1nLnQIDAQAB
-oz8wPTAdBgNVHQ4EFgQUOEfrzwBmZCNp+YOGJ1Iu/lKwyQUwDwYDVR0TBAgwBgEB
-/wIBADALBgNVHQ8EBAMCAQYwDQYJKoZIhvcNAQEMBQADggGBAOMvR0vHyDEwQYRx
-qTxKbUbGVApm560A6Du5xz7tGfMl723Y/tE7LIwgb8YutRvXVyKl2mLCEjcQszcz
-SL4wg0fDNbVHC7ogHFPFlL4tm3I8uvBElbMBZUgldOwI5P1z7Uzo1rtanfLwtpyP
-bIQtnj3WCBa+OxYYE4TmCbgNkwH0UBUGu6rsf4tX4GELFTZaosP1bAGFvEGL1411
-Q3fZpgpP1Qkrqobrb28Em7lyJJEenaC/DJdgGSQsoNgIMYx+y4Va9WpG6vo9wHkJ
-XKAIRXGO0bZ4RWq83XaQLI6zrlWBg3l977WiAmc+iCnAVYiUb3S9eRKm8fWAxYHY
-7FtoOAs2a27jWZw6YLndfi2ckt2kpB+MZWnz7zaC5HmSoW04IkWm4H/PuqkL89qZ
-cHzmHaAOFZbzQCWKCHNlUhBH48etomJJhAaN1Arbk9jnIqklVzzbyzaZBY4GDcqy
-ygbYHFvbSmYtxOK0ATDESZCRmwHJ+fRuO2G0fnb04+UpCtxBZg==
------END CERTIFICATE-----
